Improve and extend Listbox components
This commit brings enhancements to the Listbox component by introducing the 'shouldFocusOnHover' prop and refining the typing of VirtualizedListboxProps. The modifications promote better customization of the list boxes and lay groundwork for more sophisticated property handling mechanisms. Furthermore, the diffs include the addition of the 'splitProps' function to separate 'options' and 'itemLabel', thereby refining local component settings.